fix: merge Connection into FSM for better types
